extract common spec examples from IsConfigurable spec into IsConfigurableToolkit::SharedExamples, re-implement in the IsConfigurable spec
